Extract authorized vs unauthorized routes
