+++ src/Room/Messages.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useState } from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import { Box } from 'rebass'
 
-import { WebsocketContext } from 'App'
+import { WebsocketContext } from 'Router'
 
 import { MESSAGES_QUERY, MessagesQuery, Message as MessageType } from './graphql'
 import Message from './Message'
