+++ src/Router/index.ts
@@ -1,0 +1,4 @@
+import Router from './Router'
+
+export { WebsocketContext } from './Authorized'
+export default Router
